+++ src/lib/metadata.ts
@@ -3,7 +3,7 @@ import { Metadata } from 'next'
 import { IModuleInfo } from './module-info'
 
 export function makeMetaData(title: string): Metadata {
-  title = `${title} - ${SITE_NAME}`
+  title = `${title} | ${SITE_NAME}`
   return {
     metadataBase: new URL(SITE_URL),
     title,
@@ -16,7 +16,7 @@ export function makeMetaData(title: string): Metadata {
 }
 
 export function makeMetaDataFromModule(module: IModuleInfo): Metadata {
-  const title = `${module.name} - ${SITE_NAME}`
+  const title = `${module.name} | ${SITE_NAME}`
   return {
     metadataBase: new URL(SITE_URL),
     title,