+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- UI fixes. Changes to single cell module.
 - Bug fixes and updates.
 - Updated help urls.